AppPropsContext exsiting in App.tsx was the issue :D
