+++ packages/interface/src/App.tsx
@@ -5,29 +5,13 @@ import { ErrorBoundary } from 'react-error-boundary';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { MemoryRouter } from 'react-router-dom';
 
+import { AppProps, AppPropsContext } from './AppPropsContext';
 import { AppRouter } from './AppRouter';
 import { ErrorFallback } from './ErrorFallback';
 import { useCoreEvents } from './hooks/useCoreEvents';
 import './style.scss';
 
 const queryClient = new QueryClient();
-
-export const AppPropsContext = React.createContext<AppProps | null>(null);
-
-export type Platform = 'browser' | 'macOS' | 'windows' | 'linux';
-
-export interface AppProps {
-	transport: BaseTransport;
-	platform: Platform;
-	convertFileSrc: (url: string) => string;
-	openDialog: (options: { directory?: boolean }) => Promise<string | string[] | null>;
-	onClose?: () => void;
-	onMinimize?: () => void;
-	onFullscreen?: () => void;
-	onOpen?: (path: string) => void;
-	isFocused?: boolean;
-	demoMode?: boolean;
-}
 
 function RouterContainer() {
 	useCoreEvents();
