+++ packages/interface/src/screens/Content.tsx
@@ -1,4 +1,3 @@
-// import { useBridgeCommand, useBridgeQuery } from '@sd/client';
 import React from 'react';
 
 export const ContentScreen: React.FC<{}> = (props) => {
