+++ packages/interface/src/screens/Debug.tsx
@@ -2,7 +2,7 @@ import { useBridgeCommand, useBridgeQuery } from '@sd/client';
 import { Button } from '@sd/ui';
 import React, { useContext } from 'react';
 
-import { AppPropsContext } from '../App';
+import { AppPropsContext } from '../AppPropsContext';
 import CodeBlock from '../components/primitive/Codeblock';
 
 export const DebugScreen: React.FC<{}> = (props) => {
