+++ packages/interface/src/screens/Tag.tsx
@@ -9,10 +9,7 @@ export const TagScreen: React.FC<{}> = () => {
 
 	return (
 		<div className="w-full p-5">
-			<p className="px-5 py-3 mb-3 text-sm text-gray-400 rounded-md bg-gray-50 dark:text-gray-400 dark:bg-gray-600">
-				<b>Note: </b>This is a pre-alpha build of Spacedrive, many features are yet to be
-				functional.
-			</p>
+			<h1>{id}</h1>
 		</div>
 	);
 };
